Some code to find radius dependent mask to apply SGD
Does not work yet. Idea is to use the gaussian extrapolate function to find cells that are within a certain radius of the initial release point. Then also set these to TRUE. Does only work on 1 process now, otherwise you get an error on the gather to all. Also: it seems to apply a very different radius for different locations.